Create test accounts by modifying genesis.json
